Using modules in Vuex.
